+++ lib/turbostreamer/errors.rb
@@ -4,8 +4,7 @@ class TurboStreamer
   module Errors
     class MergeError < ::StandardError
       def self.build(updates)
-        message = "Can't merge #{updates.inspect} which isn't Hash or Array"
-        new(message)
+        new("Can't merge #{updates.inspect} which isn't Hash or Array")
       end
     end
   end