+++ lib/turbostreamer/handler.rb
@@ -17,8 +17,8 @@ class TurboStreamer
     
     def self.call(template, source)
       # this juggling is required to keep line numbers right in the error
-      %{__already_defined = defined?(json); json||=TurboStreamer::Template.new(self, output_buffer: output_buffer || ActionView::OutputBuffer.new); #{source}
-        json.target! unless (__already_defined && __already_defined != "method")}
+      %{__parent_json = local_assigns[:json]; json = __parent_json || TurboStreamer::Template.new(self, output_buffer: ActionView::TurboBuffer.wrap(output_buffer)); #{source}
+        json.target! unless __parent_json}
     end
     
   end